refactor: intergrate real UI card
- replace placeholder UI card with real UI card

Resolves: DCMAW-10843
