--- modules/sdk/src/main/java/uk/gov/onelogin/criorchestrator/sdk/CriOrchestrator.kt
@@ -1,5 +1,0 @@
-package uk.gov.onelogin.criorchestrator.sdk
-
-class CriOrchestrator {
-    val greeting = "HELLO WORLD"
-}
